+++ gusto_embedded/src/funcs/holidayPayPoliciesDelete.ts
@@ -21,6 +21,10 @@ import {
 import { ResponseValidationError } from "../models/errors/responsevalidationerror.js";
 import { SDKValidationError } from "../models/errors/sdkvalidationerror.js";
 import {
+  UnprocessableEntityErrorObject,
+  UnprocessableEntityErrorObject$inboundSchema,
+} from "../models/errors/unprocessableentityerrorobject.js";
+import {
   DeleteV1CompaniesCompanyUuidHolidayPayPolicyRequest,
   DeleteV1CompaniesCompanyUuidHolidayPayPolicyRequest$outboundSchema,
   DeleteV1CompaniesCompanyUuidHolidayPayPolicyResponse,
@@ -44,6 +48,7 @@ export function holidayPayPoliciesDelete(
 ): APIPromise<
   Result<
     DeleteV1CompaniesCompanyUuidHolidayPayPolicyResponse,
+    | UnprocessableEntityErrorObject
     | GustoEmbeddedError
     | ResponseValidationError
     | ConnectionError
@@ -69,6 +74,7 @@ async function $do(
   [
     Result<
       DeleteV1CompaniesCompanyUuidHolidayPayPolicyResponse,
+      | UnprocessableEntityErrorObject
       | GustoEmbeddedError
       | ResponseValidationError
       | ConnectionError
@@ -107,7 +113,7 @@ async function $do(
   );
 
   const headers = new Headers(compactMap({
-    Accept: "*/*",
+    Accept: "application/json",
     "X-Gusto-API-Version": encodeSimple(
       "X-Gusto-API-Version",
       payload["X-Gusto-API-Version"],
@@ -168,6 +174,7 @@ async function $do(
 
   const [result] = await M.match<
     DeleteV1CompaniesCompanyUuidHolidayPayPolicyResponse,
+    | UnprocessableEntityErrorObject
     | GustoEmbeddedError
     | ResponseValidationError
     | ConnectionError
@@ -181,7 +188,8 @@ async function $do(
       204,
       DeleteV1CompaniesCompanyUuidHolidayPayPolicyResponse$inboundSchema,
     ),
-    M.fail([404, 422, "4XX"]),
+    M.jsonErr([404, 422], UnprocessableEntityErrorObject$inboundSchema),
+    M.fail("4XX"),
     M.fail("5XX"),
   )(response, req, { extraFields: responseFields });
   if (!result.ok) {