+++ gusto_embedded/src/funcs/reimbursementsDeleteV1RecurringReimbursements.ts
@@ -21,6 +21,10 @@ import {
 import { ResponseValidationError } from "../models/errors/responsevalidationerror.js";
 import { SDKValidationError } from "../models/errors/sdkvalidationerror.js";
 import {
+  UnprocessableEntityErrorObject,
+  UnprocessableEntityErrorObject$inboundSchema,
+} from "../models/errors/unprocessableentityerrorobject.js";
+import {
   DeleteV1RecurringReimbursementsRequest,
   DeleteV1RecurringReimbursementsRequest$outboundSchema,
   DeleteV1RecurringReimbursementsResponse,
@@ -44,6 +48,7 @@ export function reimbursementsDeleteV1RecurringReimbursements(
 ): APIPromise<
   Result<
     DeleteV1RecurringReimbursementsResponse,
+    | UnprocessableEntityErrorObject
     | GustoEmbeddedError
     | ResponseValidationError
     | ConnectionError
@@ -69,6 +74,7 @@ async function $do(
   [
     Result<
       DeleteV1RecurringReimbursementsResponse,
+      | UnprocessableEntityErrorObject
       | GustoEmbeddedError
       | ResponseValidationError
       | ConnectionError
@@ -103,7 +109,7 @@ async function $do(
   const path = pathToFunc("/v1/recurring_reimbursements/{id}")(pathParams);
 
   const headers = new Headers(compactMap({
-    Accept: "*/*",
+    Accept: "application/json",
     "X-Gusto-API-Version": encodeSimple(
       "X-Gusto-API-Version",
       payload["X-Gusto-API-Version"],
@@ -164,6 +170,7 @@ async function $do(
 
   const [result] = await M.match<
     DeleteV1RecurringReimbursementsResponse,
+    | UnprocessableEntityErrorObject
     | GustoEmbeddedError
     | ResponseValidationError
     | ConnectionError
@@ -174,7 +181,8 @@ async function $do(
     | SDKValidationError
   >(
     M.nil(204, DeleteV1RecurringReimbursementsResponse$inboundSchema),
-    M.fail([404, "4XX"]),
+    M.jsonErr(404, UnprocessableEntityErrorObject$inboundSchema),
+    M.fail("4XX"),
     M.fail("5XX"),
   )(response, req, { extraFields: responseFields });
   if (!result.ok) {